+++ pkgs/dart_mcp_server/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add additional analytics for initialization events, and various list* method
   calls. This will help understand how the server is being used by different
   clients, and what features clients support.
+- Add a tool for running ripgrep on package dependencies.
 
 # 0.1.2 (Dart SDK 3.11.0)